Refactor in Progress
Initial work commit. Revises a lot of stuff, including gun descriptions, reloading, and chambering. Still working out the problems and currently attempting to refactor revolver code. After finishing that, will move on to changing up how in_chamber is handled so that it no longer refers to a projectile, along with changes to the chambering cycle.