Good Enough: Revolver "Fix"
Fixes revolvers a little bit from last commit. Still a lot of work to do, but I need to make some major changes to the fire cycle. That's the next step. After that I will need to implement another category of flags, similar to what TGCM uses for ammo receivers.