More WIP Progress, cleanup on a lot of code
I've made the necessary changes to in_chamber, but now I stand at a crossroads. Do I continue and adjust and fix everything else, or do I continue making changes to the backend gun cycle? The current idea is to turn in_chamber into a referenced list with an index of 1, allowing it to store both the ammo datum and a reference to a casing path for any used casings. This would allow for dynamic casings based on what ammo is actually loaded in the gun. I can't think of a way to make it work better than the current system for casings at the moment, but that may change in the future if vis_contents receives more flags. I've also finished ammo counter behavior on pickup/storage/etc.