More work on converting attachments to vs_contents
I think I've done the brunt of the work on attachment layering. The only thing remaining is convert mag overlays into vis_contents and getting rid of all traces of attachment overlay list. While I know how to do it, I am not looking forward to it as it's going to be time consuming. Lots of refacoring and moving things around on the attachment side of things.