Large re-organization, refactor of more attach code, moving things around, and other WIP changes. It's getting there
Once again we're working in progress. This time I wasn't able to compile the game for some time as I was heavy at work refactoring a lot of the attachment code, along with gun setup, lists, and various things like that. Suffice to say a lot was changed or moved around. I've organized, cleaned up, and slimmed down a lot of the attachment code as well as gun stats. Same for attachments. Everything compiles and I've  worked out some of the visual glitches I've encountered. The main problem is that I undertook a rework of how ammo magazine are shown with the gun, and right now a lot of them, including gun icons will not display correctly. It will be the next thing I finish, I think. After that more testing and on to a new revision: making sure detachable magazines can handle mixed ammo types. I already have the code, I just haven't implemented it. The difficulty lies not in firing the gun but reloading it with ammo. I may also refactor and rework handfuls at some point, but that's something to think about later down the road. For now I'm just happy I can finally test my changes after almost a week or so of just programming.