Rework of cylinder feeding/firing, cleanup, macros, etc.
Makes bolt action rifles into a subtype of rifle, a bunch more cleanup, and so forth. Made certain things into macros, like the ammo counter and rotating the cylinder. Major change is I've reworked and plugged in new mixed-ammo mechanics for revolvers. It should be fully working if not close to it. But I am committing this now because I am going to rework the mixed ammo mechanics again using a new staggered technique for layering ammo and counts, thanks to Lummox Jr. It turns out that associated lists are not as efficient and fast as I would like,  and a new approach is  warranted. So yeah, once again I have to rework a bunch of stuff. I've also noticed that ammo datums are kind of superfluous and don't serve any particular purpose any more. I would like to replace them with obj/projectile subpaths instead to basically do the same thing ammo datums do now. in_chamber will be a path to a projectile instead of ammo datum. Also made some changes to alt_click behavior, which are not fully working atm. I'll come back to those when I have mixed ammo reworked again. Still unfinished with reworking attachables, but I'll come back to that. The last step of that rework will be to make attached guns actual guns instead of pseudo-guns. Not committing map changes for gun paths, as I'll do that later when I make more changes.